// ncp-ts/src/generated/SimConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SimMode } from "./SimMode.js";

/**
 * Integration / streaming configuration.
 */
export type SimConfig = { dt_ms: number, chunk_ms: number, seed: bigint | null, mode: SimMode, duration_ms: number | null, };